Revert "Feature: args for ‘run’ command"
Cause: after discussions with team, Pliz goal is to provide a standardized way to manage projects using clearly defined tasks (via pliz.yml). So allowing to change the behavior of a task with some args can lead to mistakes that Pliz tries to limit.
If you need to pass some args to a task, just use the Docker/Docker Compose commands directly.